AR-814: Refactored ScreenLayoutRegion entity to not be under tenant scope. Updated command and fixtures
