+++ fixtures/screen.yaml
@@ -8,7 +8,7 @@ App\Entity\Tenant\Screen:
     size: <randomElement([24, 42, 55, 65])>
     resolutionWidth: 1920
     resolutionHeight: 1200
-    screenLayout: '@screen_layout_abc_*'
+    screenLayout: '@screen_layout_id*'
     tenant: '@tenant_abc'
   screen_xyz_{1..100}:
     id (unique): <ulid()>
@@ -18,5 +18,5 @@ App\Entity\Tenant\Screen:
     size: <randomElement([24, 42, 55, 65])>
     resolutionWidth: 1920
     resolutionHeight: 1200
-    screenLayout: '@screen_layout_xyz_*'
+    screenLayout: '@screen_layout_id*'
     tenant: '@tenant_xyz'
