+++ fixtures/screen_layout_region.yaml
@@ -1,12 +1,6 @@
 ---
-App\Entity\Tenant\ScreenLayoutRegions:
-  screen_layout_region_abc_{1..10}:
+App\Entity\ScreenLayoutRegions:
+  screen_layout_region_{1..10}:
     id (unique): <ulid()>
     title: <sentence(4)>
     gridArea: <randomElements(['a', 'b', 'c', 'd', 'e', 'f'])>
-    tenant: '@tenant_abc'
-  screen_layout_region_xyz_{1..10}:
-    id (unique): <ulid()>
-    title: <sentence(4)>
-    gridArea: <randomElements(['a', 'b', 'c', 'd', 'e', 'f'])>
-    tenant: '@tenant_xyz'
