+++ fixtures/template.yaml
@@ -1,16 +1,8 @@
 ---
 App\Entity\Template:
-  template_abc_{1..20}:
+  template_{1..20}:
     id (unique): <ulid()>
     title: <sentence(4)>
     description: <text()>
     icon: <imageUrl(100, 100, 'icon')>
     resources: <templateResources()>
-    tenant: '@tenant_abc'
-  template_xyz_{1..20}:
-    id (unique): <ulid()>
-    title: <sentence(4)>
-    description: <text()>
-    icon: <imageUrl(100, 100, 'icon')>
-    resources: <templateResources()>
-    tenant: '@tenant_xyz'
