+++ src/Repository/ScreenLayoutRegionsRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\Repository;
 
-use App\Entity\Tenant\ScreenLayoutRegions;
+use App\Entity\ScreenLayoutRegions;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
